move pageitemlistview to its own file, fix test
